share test type templates
